# Faster convex polygon drawing for 2D images