## Description
Replaced `grid_points_in_poly` with calls to `skimage.draw.polygon_perimeter` and `scipy.ndimage.morphology.binary_fill_holes` in convex polygon drawing step for a 2D image.
For large 2D images (~10,000 x ~10,000 pixels), this substitution can result in a function-call-to-return speedup of more than 5x (from 23.0 sec to 4.4 sec for a particular image with about 150 convex hull edges) while producing a convex hull image that is nearly identical to the image created by the current drawing routine. What follows is a comparison of the `cProfile` results and convex hull images created by the two routines.

## Testing code
To test the differences between the two drawing routines, I ran the following script to do a cProfile on the `convex_hull_image` function with and without changes to the "convex_hull.py" file in my skimage package library. Results without any changes are referred to as "latest", those with my changes are referred to as "faster". Additionally, I tested both the `offset_coordinates=True` and `offset_coordinates=False` options to `convex_hull_image`.

`import cProfile
import numpy as np
import tifffile
from PIL.Image import frombytes

from skimage.morphology import convex_hull_image


test_dir = 'C:/Users/husby036/Documents/Cprojects/test_s2s/testFiles/'


def saveImage(array, path):
    if array.dtype == np.bool:
        img = frombytes(mode='1', size=array.shape[::-1], data=np.packbits(array, axis=1))
        img.save(path)
    else:
        tifffile.imsave(path, array)
    print "'{}' saved".format(path)


function = 'convex_hull_image(image, offset_coordinates=True)'

chull_type = 'latest'

image = tifffile.imread(test_dir+'image.tif')
print("{} [{}]:\n".format(function, chull_type))
cProfile.run('image_chull = {}'.format(function))
saveImage(image_chull, test_dir+'chull_{}.tif'.format(chull_type))

debug_mask = np.zeros(image.shape, dtype=np.int8)
debug_mask[image_chull] = 1
debug_mask[image] += 2
saveImage(debug_mask, test_dir+'chull_{}_inspection.tif'.format(chull_type))`

## `convex_hull_image(image, **offset_coordinates=True**)`
### `cProfile` results
"Latest"
- [ ] [offset_latest_cprof.txt](https://github.com/scikit-image/scikit-image/files/1577130/offset_latest_cprof.txt) 
> 201 function calls in 23.216 seconds
"Faster"
- [ ] [offset_faster_cprof.txt](https://github.com/scikit-image/scikit-image/files/1577132/offset_faster_cprof.txt) 
> 908 function calls (907 primitive calls) in 4.576 seconds

### Result image comparison
_Histograms show counts of pixel values with the pixel value corresponding to each histogram bar taken to be the x-axis value at the left end of the bar._
In the upper set of images, yellow pixels (value=1) are part of the convex hull while purple pixels (value=0) are not.
The lower left image titled "Difference" is an image obtained by subtracting the upper left image from the upper right image ("faster" minus "latest"). Therefore, yellow pixels (value=1) are hull pixels in the "faster" method but not hull pixels in the "latest" method, and vice-versa for purple pixels (value=-1).
The lower right image titled "Boolean Difference" is to help spot where differences in the images are located; yellow pixels (value=1) are where differences occur.
![offset_diff](https://user-images.githubusercontent.com/20846681/34230840-2f6b9278-e59f-11e7-85c3-c1ed1851de00.png)
With the "Difference" figure zoomed to red box:
![offset_diff_zoom](https://user-images.githubusercontent.com/20846681/34230854-4898478c-e59f-11e7-9d05-b37c06f14926.png)
In the "Difference" figure, yellow and purple rectangles show one-pixel differences in the border width of the drawn hull. For the `offset_coordinates=True` option, we assume the "latest" method is most correct and want to minimize the total number of blue pixels. The total number of blue pixels is is approximately 10, which is good. Since there are approximately 10,000 yellow pixels in the "Difference" figure, we see that the "faster" method almost always draws a thicker polygon border than the "current" method.

For closer inspection of the drawing routines in relation to the source image, the following "_inspection" images were created as shown in the last lines of the above testing script.
In "*_inspection" images, yellow pixels (value=3) are pixels that are both one/True/data in the source image and are part of the drawn convex hull, light blue pixels (value=1) are pixels that are zero/False/nodata in the source image and are part of the drawn convex hull, purple pixels (value=0) are neither of these things, and **green pixels (value=2) are errors -- pixels that are one/True/data in the source image but _are not_ part of the drawn convex hull**.
![offset_latest_inspect](https://user-images.githubusercontent.com/20846681/34231871-59ef4a4a-e5a3-11e7-9a8a-1b0812fa2e9f.png)
![offset_faster_inspect](https://user-images.githubusercontent.com/20846681/34232393-b9265754-e5a5-11e7-807e-ed3cb41fb1ff.png)
Both "latest" and "faster" inspection images zoomed to the red box (only shown in the former):
![offset_latest_inspect_zoom](https://user-images.githubusercontent.com/20846681/34232410-c9f4d3a8-e5a5-11e7-8117-9a2af9cb582d.png)
![offset_faster_inspect_zoom](https://user-images.githubusercontent.com/20846681/34232413-cc9fa948-e5a5-11e7-81d1-31867f3dcefa.png)
We see that neither drawing routine (_for `offset_coordinates=True`_, but not so much for the opposite option, as seen later) has any green error pixels, which is good. We also see that, in general, the "faster" method is a bit more loose on drawing the convex hull border than the "latest" method, accounting for the significant number of pixels that are hull pixels in the "faster" method but not hull pixels in the "latest" method, as seen earlier.

## `convex_hull_image(image, **offset_coordinates=False**)`
### `cProfile` results
"Latest"
- [ ] [nooffset_latest_cprof.txt](https://github.com/scikit-image/scikit-image/files/1577288/nooffset_latest_cprof.txt)
> 197 function calls in 22.974 seconds
"Faster"
- [ ] [nooffset_faster_cprof.txt](https://github.com/scikit-image/scikit-image/files/1577294/nooffset_faster_cprof.txt)
> 896 function calls (895 primitive calls) in 4.403 seconds

### Result image comparison
![nooffset_diff](https://user-images.githubusercontent.com/20846681/34232693-1d62ebe6-e5a7-11e7-8ac6-44b3b15f1b69.png)
With the "Difference" figure zoomed to red box:
![nooffset_diff_zoom](https://user-images.githubusercontent.com/20846681/34232695-1ee5422a-e5a7-11e7-8912-5c2ef08a31a3.png)

![nooffset_latest_inspect](https://user-images.githubusercontent.com/20846681/34232708-2ad40634-e5a7-11e7-8df2-bbfe30e9c26c.png)
![nooffset_faster_inspect](https://user-images.githubusercontent.com/20846681/34233060-da77612a-e5a8-11e7-847f-43b6ee7541e2.png)
Both "latest" and "faster" inspection images zoomed to the red box (only shown in the former):
![nooffset_latest_inspect_zoom](https://user-images.githubusercontent.com/20846681/34233067-df97ebf2-e5a8-11e7-9416-80f53a58870f.png)
![nooffset_faster_inspect_zoom](https://user-images.githubusercontent.com/20846681/34233070-e1380550-e5a8-11e7-89e4-8fddc8eaaed1.png)
For `offset_coordinates=False`, we see that the "latest" method has a small but significant number of "green error pixels" (as mentioned above). While it's not obvious to me why exactly the "faster" method does not have any green error pixels, this is not unexpected as the method has been shown to draw a looser convex hull border than the "latest" method.


## Checklist
[It's fine to submit PRs which are a work in progress! But before they are merged, all PRs should provide:]
- [ ] Consider making the faster drawing routine an optional argument to `convex_hull_image`.


## For reviewers

(Don't remove the checklist below.)

- [ ] Check that the PR title is short, concise, and will make sense 1 year
  later.
- [ ] Check that new functions are imported in corresponding `__init__.py`.
- [ ] Check that new features, API changes, and deprecations are mentioned in
      `doc/release/release_dev.rst`.