@pep8speaks 2017-12-20
Hello @husby036! Thanks for submitting the PR.

 - In the file [`skimage/morphology/convex_hull.py`](https://github.com/scikit-image/scikit-image/blob/d40f773f8f02bb2be766a46c8202cd200ca3c35f/skimage/morphology/convex_hull.py), following are the PEP8 issues :

> [Line 79:80](https://github.com/scikit-image/scikit-image/blob/d40f773f8f02bb2be766a46c8202cd200ca3c35f/skimage/morphology/convex_hull.py#L79): [E501](https://duckduckgo.com/?q=pep8%20E501) line too long (83 > 79 characters)
> [Line 81:80](https://github.com/scikit-image/scikit-image/blob/d40f773f8f02bb2be766a46c8202cd200ca3c35f/skimage/morphology/convex_hull.py#L81): [E501](https://duckduckgo.com/?q=pep8%20E501) line too long (86 > 79 characters)