@soupault 2017-12-21
This PR description is just perfect :fire_engine: .
From my standpoint, the proposed modifications should be considered as an enhancement, and, therefore, scheduled for the integration.

@husby036 do you see any downsides of the `faster` method? Maybe some cases where it fails or underperforms in comparison with the `latest`?