@husby036 2017-12-21
@soupault Thank you! It's my first PR for a public repo, so I tried to put in my best effort.

I've updated the description with a test of a smaller image for comparison. The results are quite interesting -- I encourage you to take a look!

In short, I found that `faster` when run with `offset_coordinates=False` results in a convex hull image that shows no difference with `latest` and `offset_coordinates=True` (the ideal setting for `latest`) in the small image test case, and differs by a mere 27 pixels in the large image test case.
Yep, (`faster`, `offset_coordinates=False`) actually creates a better result image than (`faster`, `offset_coordinates=True`), and (without further testing) I believe it should be close enough to the ideal result in most cases to make the speedup benefit worth the slight chance of drawing a convex hull border that is 1-px too wide (from the ideal setting for `latest`) in some cases.