@jni 2017-12-25
Well, my mind is suitably boggled. I unfortunately won't have time to do a detailed review until late January (Merry Christmas, everyone! ;) but:

a) the performance benefit is substantial enough that I agree with @soupault that this should definitely be included one way or another very soon, but
b) we need to be *really* careful about the deprecation path. These single pixel changes may seem inconsequential in the overall scheme of things, but various downstream packages may depend on these things for their testing. In particular, we need to check that this new method gives *correct* results other than some floating point approximation errors. See [these comments](https://github.com/scikit-image/scikit-image/pull/2847#issuecomment-339532801) in #2847 for what kind of differences are acceptable or not. In those comments, I downplay the effect of single pixel changes, but I do that in favour of *not* changing things. When we do want to change the output, we must be much more careful.