@husby036 2017-12-25
@jni Thank you for your input, and Merry Christmas to you!

I agree 100% that the deprecation path must be analyzed in a stronger fashion than simply comparing the result images pixel-by-pixel. When I have more time, I would take a look at the differences between the `grid_points_in_poly` and `polygon_perimeter` methods to better understand the fine differences of the two drawing routines and the implications of making this substitution.

Pardon me for wanting to get the word out there that a speedup is possible, as it's greatly benefited me already. :) I'm actually using this faster polygon drawing routine for the processing of satellite imagery in a `con_**cave**__hull_image` algorithm I recently developed! I hope to add this function to the scikit-image package in the near future. :D