@jni 2017-12-25
> Pardon me for wanting to get the word out there that a speedup is possible, as it's greatly benefited me already. :)

Oh we are very happy to see this!

> I hope to add this function to the scikit-image package in the near future. :D

Sounds great!