@stefanv 2017-12-25
Also see https://github.com/scikit-image/scikit-image/pull/931