@jni 2018-05-01
Hi @husby036 et al!

Sorry for the long delay in getting back to this.

For a start, some of the tests are failing with index-out-of-bounds errors:
https://travis-ci.org/scikit-image/scikit-image/jobs/319439828#L4341

Others are failing due to the pixel-wise difference between this PR and the expected output, as we expected.

So, my suggestion:

- fix the errors. ;)
- add a keyword argument to `convex_hull_image`. Naming suggestion: `fast_drawing=None`, which gets converted to `False` by default in 0.14, and `True` from 0.16 onwards. See the [deprecation cycle](http://scikit-image.org/docs/dev/contribute.html#deprecation-cycle) guidelines from the contributor guide. After 0.16 we can deprecate the argument altogether.
- add tests that show the correct (pixel-wise) behaviour for `fast_drawing=True`, and pass `fast_drawing=False` for existing tests.

@husby036 are you up for this? Do you know how to update a PR? Let us know if you need guidance.