@husby036 2018-05-01
Hi again, @jni!

No worries about the delay.  I've been plowing through MATLAB code at my work since then... (has it really been 4 months? oh no)  This is a good excuse for a break. ;)

I am up for making the necessary changes.  Thank you for providing the link to the Travis CI build log -- I didn't know I could see that!

My hunch is that the current `offset_coordinates` keyword argument is causing most (if not all) of the index-out-of-bounds errors, given that the `skimage.draw.polygon_perimeter` method in `faster` that replaces the polygon-drawing functionality of `grid_points_in_poly` in `latest` essentially produces the desired coordinate offset in 2D; Giving `offset_coordinates=True` to `faster` essentially dilates the resulting convex polygon from what is desired, which will surely cause the array bounds error if the polygon is to share an edge with the border of the image.

I'll do more testing soon.  Still learning the whole GitHub thing, but now that I've found the nice documentation on how to run the automated testing for my modifications to skimage locally, I think I can figure it out. :)